requirements.txt
openai
langchain
pypdf
pinecone-client
tiktoken
python-dotenv
